backend/requirement.txt
Flask
pandas
scikit-learn
joblib
flask_cors
PyJWT
python-dotenv
cryptography
pycryptodome
